django_app/requirements.txt
Django==2.2.3
django-ckeditor==6.0.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
